% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escribir.R
\encoding{UTF-8}
\name{escribir}
\alias{escribir}
\title{Escritura}
\usage{
escribir(dir = c("", ""), ...)
}
\arguments{
\item{dir}{direccion}
}
\value{
Nada
}
\description{
Escribe el archivo, sea en una ruta completa o por un patron.
}
\details{
La funcion escribe de manera rapida los archivos.
}
\examples{
# example code
}
\seealso{
\code{\link{radix}}, \code{\link{leer}}, \code{\link{fwrite}}.
}
