% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer.R
\encoding{LATIN1}
\name{leer}
\alias{leer}
\title{Lectura de archivos dinamica}
\usage{
leer(arch = c(NULL, NULL), cod = "LATIN1", mens = T, ...)
}
\arguments{
\item{arch}{Es una cadena o un vector con 2 datos.
En el caso de una cadena, se especifica una ruta completa para hacer la lectura, tal cual como lo haria una funcion de lectura como vroom.
En el caso de un vector de 2 datos, se requiere el nombre del archivo en la primera posicion y en la segunda un patron con la carpeta a utilizar.}

\item{cod}{Se utiliza para especificar la codificacion del archivo, por defecto es "LATIN1".}

\item{...}{Son los parametros propios de las funciones \code{\link{read_excel}} y \code{\link{vroom}}.}
}
\value{
Marco de datos.
}
\description{
Lee de una manera dinamica, dependiendo de la extension, un archivo sin importar su formato.
}
\details{
Esta funcion se encarga de leer varios tipos de datos, la lectura es simplificada, por lo tanto su uso depende por completo de
la funcion \code{\link{radix}}, por el mimso motivo es esencial que se especifique el directorio raiz en la misama antes de usar esta funcion.
}
\examples{
# example code

}
\seealso{
\code{\link{radix}}
}
