% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.radix.R
\encoding{UTF-8}
\name{ls.radix}
\alias{ls.radix}
\title{Listar datos}
\usage{
ls.radix(tipo = "ds")
}
\arguments{
\item{tipo}{tipo de datos}
}
\value{
datos
}
\description{
Datos
}
\details{
Lista datos
}
\examples{
# example code

}
\seealso{
\code{\link{leer}}
}
