% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radix.R
\encoding{UTF-8}
\name{radix}
\alias{radix}
\title{Raiz de la libreria.}
\usage{
radix(
  cons = list(dir = "", cor = "", sql = c(host = "", puerto = "", usuario = "", contra =
    "", db = ""))
)
}
\arguments{
\item{cons}{Es una lista con los datos "dir", "cor" y "sql".}
}
\value{
Nada.
}
\description{
Genera una lista de datos (directorios, correo y conexion MySQL) para su uso interno global.
}
\details{
Es de uso meramente interno, no retorna un valor en especifico pero puede accederse al mismo a travez de otra funcion.
}
\examples{
# example code

}
\seealso{
\code{\link{leer}}
}
