% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unificar.R
\encoding{UTF-8}
\name{unificar}
\alias{unificar}
\title{Unificacion de archivo}
\usage{
unificar(dir = c("", ""), id = NULL)
}
\arguments{
\item{dir}{direccion}

\item{id}{identificador}
}
\value{
Marco de datos.
}
\description{
Unifica archivos en un dolo dataframe, incluso si tienen diferente extension.
}
\details{
Esta funcion se encarga de leer varios tipos de datos, la lectura es simplificada, por lo tanto su uso depende por completo de
la funcion \code{\link{radix}}, por el mimso motivo es esencial que se especifique el directorio raiz en la misama antes de usar esta funcion.
}
\examples{
# example code
}
\seealso{
\code{\link{radix}}, \code{\link{leer}}.
}
